IT'S NOT A WRAP

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: tpolonen <tpolonen@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/09 17:51:57 by tpolonen          #+#    #+#             */
-/*   Updated: 2021/11/21 14:35:19 by tpolonen         ###   ########.fr       */
+/*   Updated: 2021/11/23 19:14:18 by tpolonen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,5 +32,5 @@ size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 	}
 	if (dstsize > 0)
 		*d = '\0';
-	return (d - dst + ft_strlen(s));
+	return ((size_t)(d - dst) + ft_strlen(s));
 }
