+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: tpolonen <tpolonen@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/01 18:53:25 by tpolonen          #+#    #+#             */
-/*   Updated: 2021/11/14 17:20:21 by tpolonen         ###   ########.fr       */
+/*   Updated: 2021/11/23 18:19:46 by tpolonen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,11 @@
 
 char	*ft_strmapi(char const *s, char (*f) (unsigned int, char))
 {
-	char	*s2;
-	size_t	len;
-	size_t	i;
+	char			*s2;
+	unsigned int	len;
+	unsigned int	i;
 
-	len = ft_strlen(s);
+	len = (unsigned int) ft_strlen(s);
 	s2 = (char *) malloc(sizeof(char) * (len + 1));
 	if (!s2)
 		return (NULL);
